disable on android, fix getAllocatedMemorySize test to account for implementation-defined bucket_count for default containers
Summary:
This serves as a mitigation until we look further into the android crashes we
discovered in android instrumentation tests.

The bucket_count of the default constructed `unordered_map` is
implementation-defined. This diff takes that into account in the
`getAllocatedMemorySize` test.

`unordered_map|set` impl in Android NDK does not match the assumptions made in
`visitAllocationClasses`. Temporarily disable the tests with that method.

Reviewed By: mzlee

Differential Revision: D9578626

fbshipit-source-id: 355940b026f00f783bf65c2f1e10725d5aa920d3